+++ connectome_tools/s2f_recipe/existing_recipe.py
@@ -2,11 +2,15 @@
 
 This strategy takes parameters from already existing S2F recipe.
 """
+import logging
+from collections import Counter
 
 import lxml.etree as ET
 
 from connectome_tools.s2f_recipe.utils import BaseExecutor
 from connectome_tools.utils import Task
+
+L = logging.getLogger(__name__)
 
 
 class Executor(BaseExecutor):
@@ -27,28 +31,45 @@ class Executor(BaseExecutor):
         yield Task(_execute, recipe_path, task_group=__name__)
 
 
-def _execute(recipe_path):
-    # The parser is backward compatible, but it will break if both the old and the new formats
-    # are used in the same file, because the final order of the rules could be different.
-    # At the moment, it will break also if the rules in the new format contain selection attributes
-    # different from fromMType/toMType (e.g. fromEType/toEType, fromRegion/toRegion...).
-    tree = ET.parse(recipe_path)
-    # old format (circuit-documentation 0.0.19)
-    old_rules = [
-        (
-            (elem.attrib.pop("from"), elem.attrib.pop("to")),
-            {k: float(v) for k, v in elem.attrib.iteritems()},
-        )
+def _load_old_rules(tree):
+    """Load rules in the old format (circuit-documentation 0.0.19)."""
+    return [
+        # the values read will be saved unchanged to the final recipe
+        ((elem.attrib.pop("from"), elem.attrib.pop("to")), dict(elem.attrib))
         for elem in tree.findall("mTypeRule")
     ]
-    # new format (circuit-documentation 0.0.20)
-    new_rules = [
-        (
-            (elem.attrib.pop("fromMType"), elem.attrib.pop("toMType")),
-            {k: float(v) for k, v in elem.attrib.iteritems()},
-        )
+
+
+def _load_new_rules(tree):
+    """Load rules in the new format (circuit-documentation 0.0.20)."""
+    return [
+        # the values read will be saved unchanged to the final recipe
+        ((elem.attrib.pop("fromMType"), elem.attrib.pop("toMType")), dict(elem.attrib))
         for elem in tree.findall("rule")
     ]
+
+
+def _is_unique(rules):
+    counter = Counter(pathway for pathway, params in rules)
+    if len(counter) != len(rules):
+        duplicates = sorted(pathway for pathway, count in counter.items() if count > 1)
+        L.warning("Rules using the same pathway: %s", duplicates)
+        return False
+    return True
+
+
+def _execute(recipe_path):
+    """Return the rules read from an existing recipe."""
+    # The parser is backward compatible, but it will raise an exception:
+    # - if both the old and the new formats are used in the same file, because the order
+    #   of the resulting rules is not granted to be the same
+    # - if multiple rules with the same pathway exists, because they would be merged together,
+    #   regardless of any other selection attribute (e.g. fromRegion/toRegion...)
+    tree = ET.parse(recipe_path)
+    old_rules = _load_old_rules(tree)
+    new_rules = _load_new_rules(tree)
     if new_rules and old_rules:
-        raise ValueError("Rules in different formats cannot be used in the same file")
+        raise ValueError("Rules in different formats in the same file cannot be imported")
+    if not _is_unique(old_rules) or not _is_unique(new_rules):
+        raise ValueError("Rules using the same pathway cannot be imported")
     return new_rules or old_rules